Refactored Simple Geometry out of coverage
