+++ cdm/src/main/java/ucar/nc2/ft/SimpleGeometryStandardFactory.java
@@ -6,10 +6,9 @@ package ucar.nc2.ft;
 
 import ucar.nc2.constants.FeatureType;
 import ucar.nc2.dataset.NetcdfDataset;
-import ucar.nc2.ft2.coverage.adapter.SimpleGeometryCSBuilder;
-import ucar.nc2.ft2.coverage.simpgeometry.SimpleGeometryFeatureDataset;
+import ucar.nc2.ft2.simpgeometry.SimpleGeometryFeatureDataset;
+import ucar.nc2.ft2.simpgeometry.adapter.SimpleGeometryCSBuilder;
 import ucar.nc2.util.CancelTask;
-import ucar.nc2.ft2.coverage.simpgeometry.SimpleGeometryFeatureDataset;
 
 import java.io.IOException;
 import java.util.Formatter;
