+++ cdm/src/main/java/ucar/nc2/ft2/simpgeometry/CFPolygon.java
@@ -1,4 +1,4 @@
-package ucar.nc2.ft2.coverage.simpgeometry;
+package ucar.nc2.ft2.simpgeometry;
 
 import java.util.List;
 import java.io.IOException;
