+++ cdm/src/main/java/ucar/nc2/ft2/simpgeometry/SimpleGeometryCoverage.java
@@ -2,7 +2,7 @@
  * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.nc2.ft2.coverage.simpgeometry;
+package ucar.nc2.ft2.simpgeometry;
 
 import ucar.ma2.DataType;
 import ucar.ma2.InvalidRangeException;
@@ -11,8 +11,8 @@ import ucar.nc2.AttributeContainerHelper;
 import ucar.nc2.Dimension;
 import ucar.nc2.VariableSimpleIF;
 import ucar.nc2.dataset.CoordinateAxis;
+import ucar.nc2.ft2.simpgeometry.adapter.SimpleGeometryCS;
 import ucar.nc2.util.Indent;
-import ucar.nc2.ft2.coverage.adapter.SimpleGeometryCS;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
