+++ cdm/src/main/java/ucar/nc2/ft2/simpgeometry/SimpleGeometryReader.java
@@ -1,4 +1,4 @@
-package ucar.nc2.ft2.coverage.simpgeometry;
+package ucar.nc2.ft2.simpgeometry;
 
 import ucar.nc2.Attribute;
 import ucar.nc2.Variable;
