+++ cdm/src/main/java/ucar/nc2/ft2/simpgeometry/adapter/SimpleGeometryCSBuilder.java
@@ -2,7 +2,7 @@
  * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.nc2.ft2.coverage.adapter;
+package ucar.nc2.ft2.simpgeometry.adapter;
 
 import ucar.nc2.Dimension;
 import ucar.nc2.Variable;
@@ -10,11 +10,11 @@ import ucar.nc2.constants.AxisType;
 import ucar.nc2.constants.CF;
 import ucar.nc2.constants.FeatureType;
 import ucar.nc2.dataset.*;
-import ucar.nc2.ft2.coverage.simpgeometry.GeometryType;
-import ucar.nc2.ft2.coverage.simpgeometry.Line;
-import ucar.nc2.ft2.coverage.simpgeometry.Point;
-import ucar.nc2.ft2.coverage.simpgeometry.Polygon;
-import ucar.nc2.ft2.coverage.simpgeometry.SimpleGeometryReader;
+import ucar.nc2.ft2.simpgeometry.GeometryType;
+import ucar.nc2.ft2.simpgeometry.Line;
+import ucar.nc2.ft2.simpgeometry.Point;
+import ucar.nc2.ft2.simpgeometry.Polygon;
+import ucar.nc2.ft2.simpgeometry.SimpleGeometryReader;
 import ucar.unidata.geoloc.Projection;
 import ucar.unidata.geoloc.ProjectionImpl;
