+++ cdm/src/test/java/ucar/nc2/ft2/simpgeometry/TestSimpleGeometryCoverage.java
@@ -1,4 +1,4 @@
-package ucar.nc2.ft2.coverage.simpgeometry;
+package ucar.nc2.ft2.simpgeometry;
 
 import ucar.ma2.DataType;
 import ucar.ma2.InvalidRangeException;
@@ -12,12 +12,20 @@ import java.util.List;
 import org.junit.Assert;
 import org.junit.Test;
 import static org.mockito.Mockito.*;
+import static ucar.nc2.ft2.simpgeometry.GeometryType.LINE;
+import static ucar.nc2.ft2.simpgeometry.GeometryType.POINT;
+import static ucar.nc2.ft2.simpgeometry.GeometryType.POLYGON;
 import static org.mockito.BDDMockito.*;
-import ucar.nc2.ft2.coverage.adapter.SimpleGeometryCS;
 
-import static ucar.nc2.ft2.coverage.simpgeometry.GeometryType.LINE;
-import static ucar.nc2.ft2.coverage.simpgeometry.GeometryType.POINT;
-import static ucar.nc2.ft2.coverage.simpgeometry.GeometryType.POLYGON;
+import ucar.nc2.ft2.simpgeometry.CFLine;
+import ucar.nc2.ft2.simpgeometry.CFPoint;
+import ucar.nc2.ft2.simpgeometry.CFPolygon;
+import ucar.nc2.ft2.simpgeometry.GeometryType;
+import ucar.nc2.ft2.simpgeometry.Line;
+import ucar.nc2.ft2.simpgeometry.Point;
+import ucar.nc2.ft2.simpgeometry.Polygon;
+import ucar.nc2.ft2.simpgeometry.SimpleGeometryCoverage;
+import ucar.nc2.ft2.simpgeometry.adapter.SimpleGeometryCS;
 
 public class TestSimpleGeometryCoverage {
