+++ tds/src/test/java/thredds/server/ncss/dataservice/FeatureDatasetTypeTest.java
@@ -27,7 +27,7 @@ import thredds.core.TdsRequestedDataset;
 import thredds.mock.web.MockTdsContextLoader;
 import ucar.nc2.constants.FeatureType;
 import ucar.nc2.ft.FeatureDataset;
-import ucar.nc2.ft2.coverage.simpgeometry.SimpleGeometryFeatureDataset;
+import ucar.nc2.ft2.simpgeometry.SimpleGeometryFeatureDataset;
 import ucar.unidata.util.test.category.NeedsCdmUnitTest;
 
 /**
